tracejs basis of tactic monkeypatching
Introduce our own version of LibTactics that redirect "interesting"
tactic applications through the tracejs tactic plugin.

A couple of apply tactics have to be changed to sapply (or another
equivalent user-defined tactic) as we are unable to correctly redefine
the builtin apply tactic.

`make tracejs` will now run a check of coq/JsCorrectness.v with the
tracejs coq plugin enabled.