Fix issue with array assignment from within v8 library
In %MoveArrayContents, array[i] was used to assign fields of the array.
This relies on [[CanPut]] being true in order to call
[[DefineOwnProperty]] (as specified in the individual Array methods).

However, if a prototype of Array has a getter (but no setter) defined
for the desired property, [[CanPut]] will always return false.

We should therefore use Object.defineProperty() (or similar) to force
the use of [[DefineOwnProperty]].

Fixes (#27)
tests/test262/ch15/15.4/15.4.4/15.4.4.19/15.4.4.19-8-b-6.js
tests/test262/ch15/15.4/15.4.4/15.4.4.20/15.4.4.20-9-b-6.js